Fix InstrumentedHBase to time actual request time
Since operations on the HBase classes are all asynchronous, merely
timing the operation would be ineffective. Instead, we should time how
long it takes to complete the underlying request.

We do this by starting a `Timer` and attaching a `Callback` to the
request's `Deferred` result that stops the timer either when it
completes or when it raises an error.