Changed additional modules to be written in Java
In order to work a little more like the rest of Dropwizard, the modules
are now written in Java.

There is a Scala module, which applies Scala idioms on top of the Java
API.

Currently, we're missing tests and more comprehensive documentation.

The Kafka module has been temporarily removed until it can be worked in
to an idiomatic module.