Fix aggregation of docs by parent POM
The generated site contained an aggregated set of docs for all modules
in the root of the project.

It featured broken links to dependencies and gave the erroneous
impression that classes from adjacent modules were related.

This also disables the generation of test docs, as tests are not a part
of the public API.