Remove github plugin
We can't portably deploy a site to GitHub Pages using their plugin, so
lets remove it.

Ideally, we should be able to just use Maven's native Git support and
just override the branch to `gh-pages`. Unfortunately, you can't
override the branch in the POM at the moment.

For now, lets disable deployment of the site by commenting out the URL.