Extract Executor logic from ThreadPooledConsumer
The `KafkaConsumerFactory` now takes care of building the
`ExecutorService`, allowing the user to provide a custom one.

Since the provided `ExecutorService` is `Managed` externally, the
shutdown process of `ThreadPooledConsumer` no longer requires the
`ExecutorService` to be shutdown explicitly.

Because the shutdown process has changed, streams currently being
processed will not be interrupted on shutdown. Instead, they will finish
processing whatever messages are waiting and then complete.